// ui/lib/api/schemas/page.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * OpenAPI spec version: 0.0.1
 */
import type { ChapterId } from './chapterId';
import type { PageId } from './pageId';
import type { PageNodes } from './pageNodes';

export interface Page {
  chapterId?: null | ChapterId;
  /** @minimum 0 */
  height: number;
  id: PageId;
  name: string;
  /** Stacking = insertion order. Bottom-first: `source` is typically first,
   * `rendered` typically last. */
  nodes: PageNodes;
  /** @minimum 0 */
  width: number;
}